Add files via upload
🚀 Exciting News from my CodSoft Internship! 🌐✨

Thrilled to share that I've just created a dynamic and interactive To-Do list using the powerful trio of HTML, CSS, and JavaScript during my internship at CodSoft! 🚧💻

This hands-on experience has not only honed my coding skills but also taught me the art of blending aesthetics with functionality. 🎨💡

Feeling accomplished and ready to take on more challenges in the coding realm! 🔍🚀 #CodSoftInternship #HTMLCSSJavaScript #CodingAchievement #TechJourney